pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devpulse"
version = "0.1.0"
description = "Local AI-powered code review CLI — no cloud, no leaks."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "typer",
    "rich",
    "gitpython",
    "requests",
    "tomli",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]

[project.scripts]
devpulse = "devpulse.cli:app"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]